Remove special-case optimisation of ifTrue: and ifFalse:.
Like some other Smalltalks (but unlike most SOMs), SOMpp treats message sends
whose name is "ifTrue:" or "ifFalse:" as if they were boolean calls. When this
assumption is incorrect, bad things happen. In my opinion, this is not a safe
optimisation, since it's quite easy to write a program where this optimisation
causes problems. Since no other SOM that I know of does the same thing, it seems
best off to remove it.

Note that this is untested, because I don't have the Krun bits on my machine.